requirements.txt
psycopg2==2.9.4
psycopg2-binary==2.9.4
airflow-clickhouse-plugin==0.8.2
airflow-providers-clickhouse==0.0.1
apache-airflow-providers-apache-hdfs==2.2.3
apache-airflow-providers-apache-spark==3.0.0
apache-airflow-providers-apache-sqoop==3.1.1
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-docker==3.4.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-jdbc==3.2.0
apache-airflow-providers-mongo==3.1.0
apache-airflow-providers-mysql==3.2.0
apache-airflow-providers-postgres==5.2.1
apache-airflow-providers-sqlite==3.2.0
apache-airflow-providers-ssh==3.2.0
apache-airflow-providers-trino==4.3.0
acryl-datahub==0.10.4.3
acryl-datahub-airflow-plugin[plugin-v2]==0.10.4.3